Conflicts:
	docs/gotcha.txt
	docs/previousVersion.sql
	docs/templates.txt
	lib/WebGUI.pm
	lib/WebGUI/Asset.pm
	lib/WebGUI/Asset/Event.pm
	lib/WebGUI/Asset/File.pm
	lib/WebGUI/Asset/MapPoint.pm
	lib/WebGUI/Asset/RichEdit.pm
	lib/WebGUI/Asset/Sku/Product.pm
	lib/WebGUI/Asset/Snippet.pm
	lib/WebGUI/Asset/Story.pm
	lib/WebGUI/Asset/Template.pm
	lib/WebGUI/Asset/Template/TemplateToolkit.pm
	lib/WebGUI/Asset/Wobject/Calendar.pm
	lib/WebGUI/Asset/Wobject/Carousel.pm
	lib/WebGUI/Asset/Wobject/Collaboration.pm
	lib/WebGUI/Asset/Wobject/Dashboard.pm
	lib/WebGUI/Asset/Wobject/DataForm.pm
	lib/WebGUI/Asset/Wobject/Folder.pm
	lib/WebGUI/Asset/Wobject/Map.pm
	lib/WebGUI/Asset/Wobject/Search.pm
	lib/WebGUI/Asset/Wobject/Shelf.pm
	lib/WebGUI/Asset/Wobject/StockData.pm
	lib/WebGUI/Asset/Wobject/StoryTopic.pm
	lib/WebGUI/Asset/Wobject/SyndicatedContent.pm
	lib/WebGUI/Asset/Wobject/Thingy.pm
	lib/WebGUI/Asset/Wobject/WeatherData.pm
	lib/WebGUI/AssetClipboard.pm
	lib/WebGUI/AssetCollateral/DataForm/Entry.pm
	lib/WebGUI/AssetExportHtml.pm
	lib/WebGUI/AssetLineage.pm
	lib/WebGUI/AssetMetaData.pm
	lib/WebGUI/AssetTrash.pm
	lib/WebGUI/AssetVersioning.pm
	lib/WebGUI/Auth.pm
	lib/WebGUI/Cache/CHI.pm
	lib/WebGUI/Content/AssetManager.pm
	lib/WebGUI/Fork/ProgressBar.pm
	lib/WebGUI/Form/JsonTable.pm
	lib/WebGUI/Form/TimeField.pm
	lib/WebGUI/Form/Zipcode.pm
	lib/WebGUI/Group.pm
	lib/WebGUI/International.pm
	lib/WebGUI/Macro/AssetProxy.pm
	lib/WebGUI/Macro/FileUrl.pm
	lib/WebGUI/Operation/SSO.pm
	lib/WebGUI/Operation/User.pm
	lib/WebGUI/Role/Asset/Subscribable.pm
	lib/WebGUI/Shop/Cart.pm
	lib/WebGUI/Shop/Transaction.pm
	lib/WebGUI/Shop/TransactionItem.pm
	lib/WebGUI/Test.pm
	lib/WebGUI/URL/Content.pm
	lib/WebGUI/URL/Uploads.pm
	lib/WebGUI/User.pm
	lib/WebGUI/Workflow/Activity/ExtendCalendarRecurrences.pm
	lib/WebGUI/Workflow/Activity/SendNewsletters.pm
	lib/WebGUI/i18n/English/Asset.pm
	lib/WebGUI/i18n/English/WebGUI.pm
	sbin/installClass.pl
	sbin/rebuildLineage.pl
	sbin/search.pl
	sbin/testEnvironment.pl
	t/Asset/Asset.t
	t/Asset/AssetClipboard.t
	t/Asset/AssetLineage.t
	t/Asset/AssetMetaData.t
	t/Asset/Event.t
	t/Asset/File.t
	t/Asset/File/Image.t
	t/Asset/Post/notification.t
	t/Asset/Sku.t
	t/Asset/Story.t
	t/Asset/Template.t
	t/Asset/Wobject/Collaboration/templateVariables.t
	t/Asset/Wobject/Collaboration/unarchiveAll.t
	t/Asset/Wobject/Shelf.t
	t/Auth.t
	t/Macro/EditableToggle.t
	t/Macro/FilePump.t
	t/Shop/Cart.t
	t/Shop/Transaction.t
	t/Storage.t
	t/User.t
	t/Workflow.t

+++ lib/WebGUI/Workflow/Activity/ExtendCalendarRecurrences.pm
@@ -19,6 +19,7 @@ use strict;
 use base 'WebGUI::Workflow::Activity';
 use WebGUI::International;
 use WebGUI::Asset;
+use WebGUI::VersionTag;
 use DateTime;
 
 =head1 NAME
@@ -74,8 +75,8 @@ sub execute {
         return $self->COMPLETE unless $piped;
         my ( $recurId, $rest ) = split /\|/, $piped, 2;
 
-        $self->processRecurrence( $recurId, $timeLimit )
-            and $piped = $rest;
+        $self->processRecurrence( $recurId, $timeLimit );
+        $piped = $rest;
     }
 
     $instance->setScratch( recurrences => $piped );
@@ -165,20 +166,36 @@ sub processRecurrence {
     my ( $self, $recurId, $timeLimit ) = @_;
     my $eventId = $self->findLastEventId($recurId);
     my $event   = WebGUI::Asset::Event->newById( $self->session, $eventId );
+    if (! $event) {
+        $self->session->log->warn("Unable to instanciate event with assetId $eventId");
+        return 0;
+    }
+    ##Ignore assets in the trash.  Same with assets in the clipboard since they would not be
+    ##put into the clipboard.
+    if ($event->state ne 'published') {
+        return 0;
+    }
     my $recur   = $event->getRecurrence;
 
+    my $versionTag = WebGUI::VersionTag->create($self->session, {name => 'Extend Calendar Recurrence for assetId '.$event->getId, });
+    $versionTag->setWorking();
     my $start   = $event->getDateTimeStart->truncate(to => 'day');
     my $limit   = DateTime->today->add( years => 2 );
     my $end     = $event->limitedEndDate($limit);
     my $set     = $event->dateSet( $recur, $start, $end );
     my $i       = $set->iterator;
 
-    while ( my $d = $i->next ) {
-        return if ( time > $timeLimit );
+    my $time_limit = 0;
+    DATE: while ( my $d = $i->next ) {
+        if ( time > $timeLimit ) {
+            $time_limit = 1;
+            last DATE;
+        }
         $event->generateRecurrence($d);
     }
 
-    return 1;
+    $versionTag->commit;
+    return $time_limit ? 1 : 0;
 } ## end sub processRecurrence
 
 1;
